+++ x/configuration/alias.go
@@ -3,35 +3,23 @@ package configuration
 import "github.com/iov-one/iovns/x/configuration/types"
 
 // alias for types
-
-// Config aliases types.Config
-type Config = types.Config
+type (
+	Config = types.Config // Config aliases types.Config
+	Fees   = types.Fees   // Fees aliases types.Fees
+)
 
 // alias for consts
-
-// ModuleName aliases types.ModuleName
-const ModuleName = types.ModuleName
-
-// RouterKey aliases types.RouterKey
-const RouterKey = types.RouterKey
-
-// QuerierRoute aliases types.QuerierRoute
-const QuerierRoute = types.QuerierRoute
-
-// QueryConfig aliases types.QueryConfig
-const QueryConfig = types.QueryConfig
-
-// StoreKey aliases types.StoreKey
-const StoreKey = types.StoreKey
-
-const DefaultParamSpace = types.DefaultParamSpace
-
-// Type aliases
-
-// Fees aliases types.Fees
-type Fees = types.Fees
+const (
+	ModuleName        = types.ModuleName   // ModuleName aliases types.ModuleName
+	RouterKey         = types.RouterKey    // RouterKey aliases types.RouterKey
+	QuerierRoute      = types.QuerierRoute // QuerierRoute aliases types.QuerierRoute
+	QueryConfig       = types.QueryConfig  // QueryConfig aliases types.QueryConfig
+	StoreKey          = types.StoreKey     // StoreKey aliases types.StoreKey
+	DefaultParamSpace = types.DefaultParamSpace
+)
 
 // function aliases
 
-// NewFees aliases types.NewFees
-var NewFees = types.NewFees
+var (
+	NewFees = types.NewFees // NewFees aliases types.NewFees
+)