+++ x/configuration/types/config.go
@@ -1,14 +1,16 @@
 package types
 
 import (
+	"time"
+
 	sdk "github.com/cosmos/cosmos-sdk/types"
 )
 
 // Config is the configuration of the network
 type Config struct {
-	// Owners are the configuration owner, the addresses allowed to handle fees
+	// Configurer is the configuration owner, the addresses allowed to handle fees
 	// and register domains with no superuser
-	Owners []sdk.AccAddress `json:"owners"`
+	Configurer sdk.AccAddress `json:"configurer"`
 	// ValidDomain defines a regexp that determines if a domain name is valid or not
 	ValidDomain string `json:"valid_domain"`
 	// ValidName defines a regexp that determines if an account name is valid or not
@@ -18,7 +20,7 @@ type Config struct {
 	// ValidBlockchainAddress determines a regexp for a valid blockchain address
 	ValidBlockchainAddress string `json:"valid_blockchain_address"`
 	// DomainRenew defines the duration of the domain renewal period in seconds
-	DomainRenew int64 `json:"domain_renew"`
+	DomainRenew time.Duration `json:"domain_renew"`
 	// DomainGracePeriod defines the grace period for a domain deletion in seconds
-	DomainGracePeriod int64 `json:"domain_grace_period"`
+	DomainGracePeriod time.Duration `json:"domain_grace_period"`
 }