+++ x/domain/client/cli/tx.go
@@ -7,6 +7,8 @@ import (
 	"fmt"
 	"io/ioutil"
 	"os"
+	"strconv"
+	"time"
 
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/client/context"
@@ -32,7 +34,6 @@ func GetTxCmd(storeKey string, cdc *codec.Codec) *cobra.Command {
 	domainTxCmd.AddCommand(flags.PostCommands(
 		getCmdRegisterDomain(cdc),
 		getCmdAddAccountCerts(cdc),
-		getCmdFlushDomain(cdc),
 		getCmdTransferAccount(cdc),
 		getCmdTransferDomain(cdc),
 		getCmdReplaceAccountTargets(cdc),
@@ -192,38 +193,6 @@ func getCmdReplaceAccountTargets(cdc *codec.Codec) *cobra.Command {
 	cmd.Flags().String("name", "", "the name of the account whose targets you want to replace")
 	cmd.Flags().String("src", "targets.json", "the file containing the new targets in json format")
 	// return cmd
-	return cmd
-}
-
-func getCmdFlushDomain(cdc *codec.Codec) *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "flush-domain",
-		Short: "flush a domain",
-		RunE: func(cmd *cobra.Command, args []string) (err error) {
-			cliCtx := context.NewCLIContext().WithCodec(cdc)
-			inBuf := bufio.NewReader(cmd.InOrStdin())
-			txBuilder := auth.NewTxBuilderFromCLI(inBuf).WithTxEncoder(utils.GetTxEncoder(cdc))
-			// get flags
-			domain, err := cmd.Flags().GetString("domain")
-			if err != nil {
-				return
-			}
-			// build msg
-			msg := &types.MsgFlushDomain{
-				Domain: domain,
-				Owner:  cliCtx.GetFromAddress(),
-			}
-			// check if valid
-			if err = msg.ValidateBasic(); err != nil {
-				return err
-			}
-			// broadcast request
-			return utils.GenerateOrBroadcastMsgs(cliCtx, txBuilder, []sdk.Msg{msg})
-		},
-	}
-	// add flags
-	cmd.Flags().String("domain", "", "name of the domain you want to flush")
-	//
 	return cmd
 }
 
@@ -561,11 +530,15 @@ func getCmdRegisterDomain(cdc *codec.Codec) *cobra.Command {
 			if err != nil {
 				return err
 			}
-			hasSuperUser, err := cmd.Flags().GetBool("has-superuser")
+			hasSuperUserStr, err := cmd.Flags().GetString("has-superuser")
 			if err != nil {
 				return err
 			}
-			accountRenew, err := cmd.Flags().GetInt64("account-renew")
+			hasSuperUser, err := strconv.ParseBool(hasSuperUserStr)
+			if err != nil {
+				return err
+			}
+			accountRenew, err := cmd.Flags().GetDuration("account-renew")
 			if err != nil {
 				return err
 			}
@@ -584,10 +557,12 @@ func getCmdRegisterDomain(cdc *codec.Codec) *cobra.Command {
 			return utils.GenerateOrBroadcastMsgs(cliCtx, txBuilder, []sdk.Msg{msg})
 		},
 	}
+
+	defaultDuration, _ := time.ParseDuration("1h")
 	// add flags
 	cmd.Flags().String("domain", "", "name of the domain you want to register")
-	cmd.Flags().Bool("has-superuser", true, "define if this domain has a superuser or not")
-	cmd.Flags().Int64("account-renew", 10000000, "account duration in seconds before expiration")
+	cmd.Flags().String("has-superuser", "true", "define if this domain has a superuser or not")
+	cmd.Flags().Duration("account-renew", defaultDuration, "account duration in seconds before expiration")
 	return cmd
 }