+++ x/domain/handler_test.go
@@ -797,14 +797,14 @@ func Test_handlerMsgRenewAccount(t *testing.T) {
 				// set mock domain
 				k.CreateDomain(ctx, types.Domain{
 					Name:         "test",
-					AccountRenew: 100,
+					AccountRenew: 1000 * time.Second,
 					Admin:        bobKey.GetAddress(),
 				})
 				// set mock account
 				k.CreateAccount(ctx, types.Account{
 					Domain:     "test",
 					Name:       "test",
-					ValidUntil: 1000,
+					ValidUntil: iovns.TimeToSeconds(time.Unix(1000, 0)),
 					Owner:      aliceKey.GetAddress(),
 				})
 			},
@@ -819,8 +819,9 @@ func Test_handlerMsgRenewAccount(t *testing.T) {
 			},
 			AfterTest: func(t *testing.T, k keeper.Keeper, ctx sdk.Context, mocks *keeper.Mocks) {
 				account, _ := k.GetAccount(ctx, "test", "test")
-				if account.ValidUntil != 1100 {
-					t.Fatalf("handlerMsgRenewAccount() expected 1100, got: %d", account.ValidUntil)
+				want := iovns.TimeToSeconds(time.Unix(1000, 0).Add(1000 * time.Second))
+				if account.ValidUntil != want {
+					t.Fatalf("handlerMsgRenewAccount() want: %d, got: %d", want, account.ValidUntil)
 				}
 			},
 		},
@@ -1539,6 +1540,7 @@ func Test_handleMsgDomainDelete(t *testing.T) {
 			AfterTest: nil,
 		},
 		"fail domain admin does not match msg owner": {
+			BeforeTestBlockTime: 0,
 			BeforeTest: func(t *testing.T, k keeper.Keeper, ctx sdk.Context, mocks *keeper.Mocks) {
 				setConfig := getConfigSetter(k.ConfigurationKeeper).SetConfig
 				setConfig(ctx, configuration.Config{
@@ -1553,6 +1555,7 @@ func Test_handleMsgDomainDelete(t *testing.T) {
 					Broker:       nil,
 				})
 			},
+			TestBlockTime: 1,
 			Test: func(t *testing.T, k keeper.Keeper, ctx sdk.Context, mocks *keeper.Mocks) {
 				_, err := handlerMsgDeleteDomain(ctx, k, &types.MsgDeleteDomain{
 					Domain: "test",
@@ -1564,7 +1567,115 @@ func Test_handleMsgDomainDelete(t *testing.T) {
 			},
 			AfterTest: nil,
 		},
-		"success": {
+		"fail domain grace period not over": {
+			BeforeTestBlockTime: 0,
+			BeforeTest: func(t *testing.T, k keeper.Keeper, ctx sdk.Context, mocks *keeper.Mocks) {
+				setConfig := getConfigSetter(k.ConfigurationKeeper).SetConfig
+				setConfig(ctx, configuration.Config{
+					DomainGracePeriod: 5,
+				})
+				k.CreateDomain(ctx, types.Domain{
+					Name:         "test",
+					Admin:        bobKey.GetAddress(),
+					ValidUntil:   3,
+					HasSuperuser: true,
+					AccountRenew: 0,
+					Broker:       nil,
+				})
+			},
+			TestBlockTime: 3,
+			Test: func(t *testing.T, k keeper.Keeper, ctx sdk.Context, mocks *keeper.Mocks) {
+				_, err := handlerMsgDeleteDomain(ctx, k, &types.MsgDeleteDomain{
+					Domain: "test",
+					Owner:  aliceKey.GetAddress(),
+				})
+				if !errors.Is(err, types.ErrUnauthorized) {
+					t.Fatalf("handlerMsgDeleteDomain() expected error: %s, got: %s", types.ErrUnauthorized, err)
+				}
+			},
+			AfterTest: nil,
+		},
+		"success domain grace period over": {
+			BeforeTestBlockTime: 0,
+			BeforeTest: func(t *testing.T, k keeper.Keeper, ctx sdk.Context, mocks *keeper.Mocks) {
+				setConfig := getConfigSetter(k.ConfigurationKeeper).SetConfig
+				setConfig(ctx, configuration.Config{
+					DomainGracePeriod: 5,
+				})
+				k.CreateDomain(ctx, types.Domain{
+					Name:         "test",
+					Admin:        bobKey.GetAddress(),
+					ValidUntil:   4,
+					HasSuperuser: true,
+					AccountRenew: 0,
+					Broker:       nil,
+				})
+			},
+			TestBlockTime: 10,
+			Test: func(t *testing.T, k keeper.Keeper, ctx sdk.Context, mocks *keeper.Mocks) {
+				_, err := handlerMsgDeleteDomain(ctx, k, &types.MsgDeleteDomain{
+					Domain: "test",
+					Owner:  aliceKey.GetAddress(),
+				})
+				if err != nil {
+					t.Fatalf("handlerMsgDeleteDomain() got error: %s", err)
+				}
+			},
+			AfterTest: nil,
+		},
+		"success owner can delete one of the domains after one expires and deleted": {
+			BeforeTestBlockTime: 1589826438,
+			BeforeTest: func(t *testing.T, k keeper.Keeper, ctx sdk.Context, mocks *keeper.Mocks) {
+				setConfig := getConfigSetter(k.ConfigurationKeeper).SetConfig
+				setConfig(ctx, configuration.Config{
+					DomainGracePeriod: 1,
+				})
+				k.CreateDomain(ctx, types.Domain{
+					Name:         "test1",
+					Admin:        bobKey.GetAddress(),
+					ValidUntil:   1589826439,
+					HasSuperuser: true,
+					AccountRenew: 0,
+					Broker:       nil,
+				})
+				k.CreateDomain(ctx, types.Domain{
+					Name:         "test2",
+					Admin:        bobKey.GetAddress(),
+					ValidUntil:   1589828251,
+					HasSuperuser: true,
+					AccountRenew: 0,
+					Broker:       nil,
+				})
+			},
+			TestBlockTime: 1589826441,
+			Test: func(t *testing.T, k keeper.Keeper, ctx sdk.Context, mocks *keeper.Mocks) {
+				// another user can delete expired domain
+				_, err := handlerMsgDeleteDomain(ctx, k, &types.MsgDeleteDomain{
+					Domain: "test1",
+					Owner:  aliceKey.GetAddress(),
+				})
+				if err != nil {
+					t.Fatalf("handlerMsgDeleteDomain() got error: %s", err)
+				}
+				_, err = handlerMsgDeleteDomain(ctx, k, &types.MsgDeleteDomain{
+					Domain: "test2",
+					Owner:  aliceKey.GetAddress(),
+				})
+				if !errors.Is(err, types.ErrUnauthorized) {
+					t.Fatalf("handlerMsgDeleteDomain() expected error: %s, got: %s", types.ErrUnauthorized, err)
+				}
+				_, err = handlerMsgDeleteDomain(ctx, k, &types.MsgDeleteDomain{
+					Domain: "test2",
+					Owner:  bobKey.GetAddress(),
+				})
+				if err != nil {
+					t.Fatalf("handlerMsgDeleteDomain() got error: %s", err)
+				}
+			},
+			AfterTest: nil,
+		},
+		"success owner can delete their domain before grace period": {
+			BeforeTestBlockTime: 0,
 			BeforeTest: func(t *testing.T, k keeper.Keeper, ctx sdk.Context, mocks *keeper.Mocks) {
 				setConfig := getConfigSetter(k.ConfigurationKeeper).SetConfig
 				setConfig(ctx, configuration.Config{
@@ -1579,19 +1690,8 @@ func Test_handleMsgDomainDelete(t *testing.T) {
 					AccountRenew: 0,
 					Broker:       nil,
 				})
-				// add two accounts
-				k.CreateAccount(ctx, types.Account{
-					Domain: "test",
-					Name:   "1",
-					Owner:  bobKey.GetAddress(),
-				})
-				// add two accounts
-				k.CreateAccount(ctx, types.Account{
-					Domain: "test",
-					Name:   "2",
-					Owner:  bobKey.GetAddress(),
-				})
 			},
+			TestBlockTime: 4,
 			Test: func(t *testing.T, k keeper.Keeper, ctx sdk.Context, mocks *keeper.Mocks) {
 				_, err := handlerMsgDeleteDomain(ctx, k, &types.MsgDeleteDomain{
 					Domain: "test",
@@ -1672,124 +1772,6 @@ func Test_handleMsgDomainDelete(t *testing.T) {
 	runTests(t, cases)
 }
 
-func Test_handlerMsgFlushDomain(t *testing.T) {
-	cases := map[string]subTest{
-		"domain does not exist": {
-			BeforeTest: nil,
-			Test: func(t *testing.T, k keeper.Keeper, ctx sdk.Context, mocks *keeper.Mocks) {
-				_, err := handlerMsgFlushDomain(ctx, k, &types.MsgFlushDomain{
-					Domain: "does not exist",
-					Owner:  nil,
-				})
-				if !errors.Is(err, types.ErrDomainDoesNotExist) {
-					t.Fatalf("handlerMsgFlushDomain() expected error: %s, got: %s", types.ErrDomainDoesNotExist, err)
-				}
-			},
-			AfterTest: nil,
-		},
-		"domain has not superuser": {
-			BeforeTest: func(t *testing.T, k keeper.Keeper, ctx sdk.Context, mocks *keeper.Mocks) {
-				k.CreateDomain(ctx, types.Domain{
-					Name:         "test",
-					Admin:        iovns.ZeroAddress,
-					ValidUntil:   0,
-					HasSuperuser: false,
-					AccountRenew: 0,
-					Broker:       nil,
-				})
-			},
-			Test: func(t *testing.T, k keeper.Keeper, ctx sdk.Context, mocks *keeper.Mocks) {
-				_, err := handlerMsgFlushDomain(ctx, k, &types.MsgFlushDomain{
-					Domain: "test",
-					Owner:  nil,
-				})
-				if !errors.Is(err, types.ErrUnauthorized) {
-					t.Fatalf("handlerMsgFlushDomain() expected error: %s, got: %s", types.ErrUnauthorized, err)
-				}
-			},
-			AfterTest: nil,
-		},
-		"msg owner is not domain admin": {
-			BeforeTest: func(t *testing.T, k keeper.Keeper, ctx sdk.Context, mocks *keeper.Mocks) {
-				k.CreateDomain(ctx, types.Domain{
-					Name:         "test",
-					Admin:        aliceKey.GetAddress(),
-					ValidUntil:   0,
-					HasSuperuser: true,
-					AccountRenew: 0,
-					Broker:       nil,
-				})
-			},
-			Test: func(t *testing.T, k keeper.Keeper, ctx sdk.Context, mocks *keeper.Mocks) {
-				_, err := handlerMsgFlushDomain(ctx, k, &types.MsgFlushDomain{
-					Domain: "test",
-					Owner:  bobKey.GetAddress(),
-				})
-				if !errors.Is(err, types.ErrUnauthorized) {
-					t.Fatalf("handlerMsgFlushDomain() expected error: %s, got: %s", types.ErrUnauthorized, err)
-				}
-			},
-			AfterTest: nil,
-		},
-		"success": {
-			BeforeTest: func(t *testing.T, k keeper.Keeper, ctx sdk.Context, mocks *keeper.Mocks) {
-				// set domain
-				k.CreateDomain(ctx, types.Domain{
-					Name:         "test",
-					Admin:        aliceKey.GetAddress(),
-					ValidUntil:   0,
-					HasSuperuser: true,
-					AccountRenew: 0,
-					Broker:       nil,
-				})
-				// add empty account 1
-				k.CreateAccount(ctx, types.Account{
-					Domain: "test",
-					Name:   "",
-					Owner:  aliceKey.GetAddress(),
-				})
-				// add account 2
-				k.CreateAccount(ctx, types.Account{
-					Domain: "test",
-					Name:   "1",
-					Owner:  bobKey.GetAddress(),
-				})
-				// add account 2
-				k.CreateAccount(ctx, types.Account{
-					Domain: "test",
-					Name:   "2",
-					Owner:  bobKey.GetAddress(),
-				})
-			},
-			Test: func(t *testing.T, k keeper.Keeper, ctx sdk.Context, mocks *keeper.Mocks) {
-				_, err := handlerMsgFlushDomain(ctx, k, &types.MsgFlushDomain{
-					Domain: "test",
-					Owner:  aliceKey.GetAddress(),
-				})
-				if err != nil {
-					t.Fatalf("handlerMsgFlushDomain() got error: %s", err)
-				}
-			},
-			AfterTest: func(t *testing.T, k keeper.Keeper, ctx sdk.Context, mocks *keeper.Mocks) {
-				var exists bool
-				_, exists = k.GetAccount(ctx, "test", "")
-				if !exists {
-					t.Fatalf("handlerMsgFlushDomain() empty account was deleted")
-				}
-				_, exists = k.GetAccount(ctx, "test", "1")
-				if exists {
-					t.Fatalf("handlerMsgFlushDomain() account 1 was not deleted")
-				}
-				_, exists = k.GetAccount(ctx, "test", "2")
-				if exists {
-					t.Fatalf("handlerMsgFlushDomain() account 2 was not deleted")
-				}
-			},
-		},
-	}
-	runTests(t, cases)
-}
-
 func TestHandleMsgRegisterDomain(t *testing.T) {
 	testCases := map[string]subTest{
 		"success": {
@@ -1797,7 +1779,7 @@ func TestHandleMsgRegisterDomain(t *testing.T) {
 				configSetter := getConfigSetter(k.ConfigurationKeeper)
 				// set config
 				configSetter.SetConfig(ctx, configuration.Config{
-					Owners:      []sdk.AccAddress{aliceKey.GetAddress()},
+					Configurer:  aliceKey.GetAddress(),
 					ValidDomain: "^(.*?)?",
 				})
 			},
@@ -1888,7 +1870,7 @@ func TestHandleMsgRegisterDomain(t *testing.T) {
 			BeforeTest: func(t *testing.T, k keeper.Keeper, ctx sdk.Context, mocks *keeper.Mocks) {
 				// add config with owner
 				config := configuration.Config{
-					Owners:                 []sdk.AccAddress{aliceKey.GetAddress()},
+					Configurer:             aliceKey.GetAddress(),
 					ValidDomain:            "^(.*?)?",
 					ValidName:              "",
 					ValidBlockchainID:      "",
@@ -1937,7 +1919,7 @@ func Test_handlerDomainRenew(t *testing.T) {
 				// add config
 				setConfig := getConfigSetter(k.ConfigurationKeeper).SetConfig
 				setConfig(ctx, configuration.Config{
-					DomainRenew: 1,
+					DomainRenew: 1 * time.Second,
 				})
 				// add domain
 				k.CreateDomain(ctx, types.Domain{
@@ -2124,4 +2106,44 @@ func Test_handlerMsgTransferDomain(t *testing.T) {
 	}
 
 	runTests(t, cases)
+}
+
+func Test_validateBlockchainTargets(t *testing.T) {
+	type args struct {
+		targets []types.BlockchainAddress
+		conf    configuration.Config
+	}
+	tests := []struct {
+		name    string
+		args    args
+		wantErr bool
+	}{
+		{
+			name: "duplicate blockchain target",
+			args: args{
+				targets: []types.BlockchainAddress{
+					{
+						ID:      "duplicate",
+						Address: "does not matter",
+					},
+					{
+						ID:      "duplicate",
+						Address: "does not matter",
+					},
+				},
+				conf: configuration.Config{
+					ValidBlockchainID:      regexMatchAll,
+					ValidBlockchainAddress: regexMatchAll,
+				},
+			},
+			wantErr: true,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			if err := validateBlockchainTargets(tt.args.targets, tt.args.conf); (err != nil) != tt.wantErr {
+				t.Errorf("validateBlockchainTargets() error = %v, wantErr %v", err, tt.wantErr)
+			}
+		})
+	}
 }