+++ x/domain/keeper/fees.go
@@ -1,10 +1,12 @@
 package keeper
 
 import (
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/cosmos/cosmos-sdk/x/distribution"
-	"github.com/iov-one/iovns/x/domain/types"
 	"log"
+
+	"github.com/cosmos/cosmos-sdk/x/auth"
+
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/iov-one/iovns/x/domain/types"
 )
 
 // CollectFees collects the fees of a msg and sends them
@@ -23,8 +25,6 @@ func (k Keeper) CollectFees(ctx sdk.Context, msg sdk.Msg, addr sdk.AccAddress) e
 	case *types.MsgDeleteAccount:
 		level = len(msg.Domain)
 	case *types.MsgDeleteDomain:
-		level = len(msg.Domain)
-	case *types.MsgFlushDomain:
 		level = len(msg.Domain)
 	case *types.MsgRegisterAccount:
 		level = len(msg.Domain)
@@ -45,5 +45,5 @@ func (k Keeper) CollectFees(ctx sdk.Context, msg sdk.Msg, addr sdk.AccAddress) e
 		return nil
 	}
 	// transfer fee to distribution
-	return k.SupplyKeeper.SendCoinsFromAccountToModule(ctx, addr, distribution.ModuleName, sdk.NewCoins(fee))
+	return k.SupplyKeeper.SendCoinsFromAccountToModule(ctx, addr, auth.FeeCollectorName, sdk.NewCoins(fee))
 }